Trending style transfered to separated Styled Component
